Added usort+black format enforcer

+++ database.py
@@ -3,11 +3,10 @@
 # Copyright (c) 2022-2023 scmanjarrez. All rights reserved.
 # This work is licensed under the terms of the MIT license.
 
+import sqlite3 as sql
 from contextlib import closing
 
-import sqlite3 as sql
-
-DB = 'crtm.db'
+DB = "crtm.db"
 
 
 def setup_db():
@@ -43,32 +42,27 @@ def cached(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
             cur.execute(
-                'SELECT EXISTS ('
-                'SELECT 1 '
-                'FROM users '
-                'WHERE uid = ?'
-                ')',
-                [uid])
+                "SELECT EXISTS ("
+                "SELECT 1 "
+                "FROM users "
+                "WHERE uid = ?"
+                ")",
+                [uid],
+            )
             return cur.fetchone()[0]
 
 
 def add_user(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
-            cur.execute(
-                'INSERT INTO users (uid) '
-                'VALUES (?)',
-                [uid])
+            cur.execute("INSERT INTO users (uid) VALUES (?)", [uid])
             db.commit()
 
 
 def del_user(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
-            cur.execute(
-                'DELETE FROM users '
-                'WHERE uid = ?',
-                [uid])
+            cur.execute("DELETE FROM users WHERE uid = ?", [uid])
             db.commit()
 
 
@@ -76,10 +70,11 @@ def favorites(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
             cur.execute(
-                'SELECT type, stop_id, stop '
-                'FROM favorites '
-                'WHERE uid = ?',
-                [uid])
+                "SELECT type, stop_id, stop "
+                "FROM favorites "
+                "WHERE uid = ?",
+                [uid],
+            )
             return cur.fetchall()
 
 
@@ -87,12 +82,13 @@ def favorite_cached(uid, transport, stop_id):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
             cur.execute(
-                'SELECT EXISTS ('
-                'SELECT 1 '
-                'FROM favorites '
-                'WHERE uid = ? AND type = ? AND stop_id = ?'
-                ')',
-                [uid, transport, stop_id])
+                "SELECT EXISTS ("
+                "SELECT 1 "
+                "FROM favorites "
+                "WHERE uid = ? AND type = ? AND stop_id = ?"
+                ")",
+                [uid, transport, stop_id],
+            )
             return cur.fetchone()[0]
 
 
@@ -100,15 +96,15 @@ def add_favorite(uid, transport, stop_id, stop):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
             cur.execute(
-                'INSERT OR IGNORE INTO transports '
-                '(type) '
-                'VALUES (?)',
-                [transport])
+                "INSERT OR IGNORE INTO transports (type) VALUES (?)",
+                [transport],
+            )
             cur.execute(
-                'INSERT OR IGNORE INTO favorites '
-                '(uid, type, stop_id, stop) '
-                'VALUES (?, ?, ?, ?)',
-                [uid, transport, stop_id, stop])
+                "INSERT OR IGNORE INTO favorites "
+                "(uid, type, stop_id, stop) "
+                "VALUES (?, ?, ?, ?)",
+                [uid, transport, stop_id, stop],
+            )
             db.commit()
 
 
@@ -116,10 +112,11 @@ def rename_favorite(uid, transport, stop_id, stop):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
             cur.execute(
-                'UPDATE favorites '
-                'SET stop = ? '
-                'WHERE uid = ? AND type = ? AND stop_id = ?',
-                [stop, uid, transport, stop_id])
+                "UPDATE favorites "
+                "SET stop = ? "
+                "WHERE uid = ? AND type = ? AND stop_id = ?",
+                [stop, uid, transport, stop_id],
+            )
             db.commit()
 
 
@@ -127,31 +124,24 @@ def del_favorite(uid, transport, stop_id):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
             cur.execute(
-                'DELETE FROM favorites '
-                'WHERE uid = ? AND type = ? AND stop_id = ?',
-                [uid, transport, stop_id])
+                "DELETE FROM favorites "
+                "WHERE uid = ? AND type = ? AND stop_id = ?",
+                [uid, transport, stop_id],
+            )
             db.commit()
 
 
 def card(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
-            cur.execute(
-                'SELECT card '
-                'FROM users '
-                'WHERE uid = ?',
-                [uid])
+            cur.execute("SELECT card FROM users WHERE uid = ?", [uid])
             return cur.fetchone()[0]
 
 
 def save_card(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
-            cur.execute(
-                'SELECT save '
-                'FROM users '
-                'WHERE uid = ?',
-                [uid])
+            cur.execute("SELECT save FROM users WHERE uid = ?", [uid])
             return cur.fetchone()[0]
 
 
@@ -159,10 +149,8 @@ def toggle_card(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
             cur.execute(
-                'UPDATE users '
-                'SET save = NOT save '
-                'WHERE uid = ?',
-                [uid])
+                "UPDATE users SET save = NOT save WHERE uid = ?", [uid]
+            )
             db.commit()
 
 
@@ -170,33 +158,21 @@ def add_card(uid, cardn):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
             cur.execute(
-                'UPDATE users '
-                'SET card = ? '
-                'WHERE uid = ?',
-                [cardn, uid])
+                "UPDATE users SET card = ? WHERE uid = ?", [cardn, uid]
+            )
             db.commit()
 
 
 def del_card(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
-            cur.execute(
-                'UPDATE users '
-                'SET card = NULL '
-                'WHERE uid = ?',
-                [uid])
+            cur.execute("UPDATE users SET card = NULL WHERE uid = ?", [uid])
             db.commit()
 
 
 def del_data(uid):
     with closing(sql.connect(DB)) as db:
         with closing(db.cursor()) as cur:
-            cur.execute(
-                'DELETE FROM favorites '
-                'WHERE uid = ?',
-                [uid])
-            cur.execute(
-                'DELETE FROM users '
-                'WHERE uid = ?',
-                [uid])
+            cur.execute("DELETE FROM favorites WHERE uid = ?", [uid])
+            cur.execute("DELETE FROM users WHERE uid = ?", [uid])
             db.commit()
